requirements.txt
Django==5.0
python-telegram-bot==13.15
environs==9.5.0
